fix: 멤버인터페이스삭제
